requirements.txt
jupyter
ipykernel
tensorflow-gpu==1.6.0
scipy
tqdm
Pillow